备份老的swipe 换上新的~
