+++ components.json
@@ -30,9 +30,7 @@
   "flex-item": "./src/components/flex-item/index.js",
   "icon": "./src/components/icon/index.js",
   "swipe": "./src/components/swipe/index.js",
-  "swipe2": "./src/components/swipe2/index.js",
-  "swipe3": "./src/components/swipe3/index.js",
-  "swipe4": "./src/components/swipe4/index.js",
+  "swipeplus": "./src/components/swipeplus/index.js",
   "swipe-item": "./src/components/swipe-item/index.js",
   "swipe-indicator": "./src/components/swipe-indicator/index.js",
   "popup": "./src/components/popup/index.js",
