+++ src/components/swipe2_tmp/index.js
