--- src/components/swipe3/index.js
@@ -1,1 +1,0 @@
-export { default } from './swipe3.vue'
