--- src/components/swipe4/index.js
@@ -1,1 +1,0 @@
-export { default } from './swipe4.vue'
