+++ src/components/swipe_old/index.js
@@ -1,0 +1,1 @@
+export { default } from './swipe.vue'
